+++ docs/conf.py
@@ -33,6 +33,7 @@ extensions = [
     'sphinxarg.ext',
     'IPython.sphinxext.ipython_console_highlighting',
     'IPython.sphinxext.ipython_directive',
+    'sphinx_click',
 ]
 
 # extension configurations