+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file, and this pr
 * Fixed bug in intermediate variable preprocessor for newer versions of Python
 * Added a context manager for checking code against a reference implementation from the same
   notebook
+* Added a CLI for interacting with PyBryt
 
 ## 0.0.5 - 2021-05-04